MPFR 3.1.1 tag.

+++ tests/tpow.c
@@ -20,8 +20,11 @@ along with the GNU MPFR Library; see the file COPYING.LESSER.  If not, see
 http://www.gnu.org/licenses/ or write to the Free Software Foundation, Inc.,
 51 Franklin St, Fifth Floor, Boston, MA 02110-1301, USA. */
 
+#include <stdio.h>
+#include <stdlib.h>
 #include <float.h>
 #include <math.h>
+#include <limits.h>
 
 #include "mpfr-test.h"
 
@@ -786,8 +789,7 @@ particular_cases (void)
           {
             printf ("Error in mpfr_pow for (%s)^(%s) (%d,%d):\n"
                     "Flags = %u instead of expected %u\n",
-                    name[i], name[j], i, j,
-                    (unsigned int) __gmpfr_flags, f[i][j]);
+                    name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
             mpfr_dump (r);
             error = 1;
           }
@@ -816,8 +818,7 @@ particular_cases (void)
               {
                 printf ("Error in mpfr_pow_z for (%s)^(%s) (%d,%d):\n"
                         "Flags = %u instead of expected %u\n",
-                        name[i], name[j], i, j,
-                        (unsigned int) __gmpfr_flags, f[i][j]);
+                        name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
                 mpfr_dump (r);
                 error = 1;
               }
@@ -842,8 +843,7 @@ particular_cases (void)
               {
                 printf ("Error in mpfr_pow_si for (%s)^(%s) (%d,%d):\n"
                         "Flags = %u instead of expected %u\n",
-                        name[i], name[j], i, j,
-                        (unsigned int) __gmpfr_flags, f[i][j]);
+                        name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
                 mpfr_dump (r);
                 error = 1;
               }
@@ -869,8 +869,7 @@ particular_cases (void)
                   {
                     printf ("Error in mpfr_pow_ui for (%s)^(%s) (%d,%d):\n"
                             "Flags = %u instead of expected %u\n",
-                            name[i], name[j], i, j,
-                            (unsigned int) __gmpfr_flags, f[i][j]);
+                            name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
                     mpfr_dump (r);
                     error = 1;
                   }
@@ -900,8 +899,7 @@ particular_cases (void)
               {
                 printf ("Error in mpfr_ui_pow for (%s)^(%s) (%d,%d):\n"
                         "Flags = %u instead of expected %u\n",
-                        name[i], name[j], i, j,
-                        (unsigned int) __gmpfr_flags, f[i][j]);
+                        name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
                 mpfr_dump (r);
                 error = 1;
               }
@@ -1111,7 +1109,7 @@ overflows2 (void)
       if (__gmpfr_flags != (MPFR_FLAGS_OVERFLOW | MPFR_FLAGS_INEXACT))
         {
           printf ("Error in overflows2 (e = %d): bad flags (%u)\n",
-                  e, (unsigned int) __gmpfr_flags);
+                  e, __gmpfr_flags);
           exit (1);
         }
     }
@@ -1160,7 +1158,7 @@ overflows3 (void)
           if (__gmpfr_flags != (MPFR_FLAGS_OVERFLOW | MPFR_FLAGS_INEXACT))
             {
               printf ("Error in overflows3 (RNDN, i = %d): bad flags (%u)\n",
-                      i, (unsigned int) __gmpfr_flags);
+                      i, __gmpfr_flags);
               exit (1);
             }
 
@@ -1290,8 +1288,7 @@ bug20071104 (void)
     }
   if (__gmpfr_flags != (MPFR_FLAGS_OVERFLOW | MPFR_FLAGS_INEXACT))
     {
-      printf ("Error in bug20071104: bad flags (%u)\n",
-              (unsigned int) __gmpfr_flags);
+      printf ("Error in bug20071104: bad flags (%u)\n", __gmpfr_flags);
       exit (1);
     }
   mpfr_clears (x, y, z, (mpfr_ptr) 0);
